Verificación, Corrección y Actualización de ABMs
Tareas #39 #18 #15 #16 #14 : Verificación, Corrección y Actualización de
CSS y Validaciones en el ABM de Clientes, Productos, Departamento,
Localidades, Usuarios, Funcionarios, etc.

+++ ventas/db/migrate/20130527013334_create_productos.rb
@@ -5,7 +5,7 @@ class CreateProductos < ActiveRecord::Migration
       t.string :descripcion, :limit => 30,  :null => false
       t.integer :cant_minima , :limit => 10,  :null => false
       t.integer :cant_optima , :limit => 10,  :null => false
-      t.decimal :precio_unitario, :limit => 10,  :null => false
+      t.integer :precio_unitario, :limit => 10,  :null => false
       t.integer :porcentaje , :limit => 2,  :null => false
     end
   end
